Compatible toy account (#3266)
# Motivation
We wish to add tests that ensure that account details remain as
expected. The recently added API method `get_test_account()` provides
useful data, but it is unrefined. In particular it has no `.did` file
entry so the response is hard to parse, and it is hard to compare
results with those from `get_account()`.

# Changes
- Make the `get_test_account` response type the same as that of
`get_account`. The type change wraps the response in an opt that is None
if the account is not there, and provides the account data in a stable
format that doesn't depend on details of the underlying `Account` type.
- Add an entry in the `nns-dapp` `.did` file for `get_test_account`.
Note: `get_test_account()` has two protections against misuse in
production: It is present only in test builds and it may only be called
by a canister controller.

# Tests
- There is a test that verifies that test accounts can be read with the
expected format and values.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>