Fix and simplify SetDissolveDelay.svelte (#3292)
# Motivation

`SetDissolveDelay.svelte` has several bugs:
1. Error messages are only updated on text input and not on slider
input.
2. When the current dissolve delay is not a whole number of days,
increasing it by less than a day to a whole number of days is allowed.
But we still show an error message that it's not allowed even though the
button is enabled.
3. The slider has second granularity but we only display days so we
might display X days but that will not actually be the dissolve delay
you get.

It is also overly complicated because:
1. The slider has granularity in seconds while the text input has
granularity in days and both numbers need to be kept in sync.
2. Disabling the button has logic based on seconds while the error
message has separate logic based on days.
3. There is a `minDelayInSeconds` prop which is not actually the minimum
allowed value but rather all allowed values are strictly greater than
this value.


# Changes

1. Use days for all logic. And only convert back to seconds (as a whole
multiple of the number of seconds in a day) to export the output prop.
2. Determine whether the button is disable based on the error message
rather than separate logic.
3. Use the `neuronDissolveDelaySeconds` for the lower bound and remove
the redundant `minDelayInSeconds` prop.
4. Update the error message also on slider input.
5. Update the tests to use days for the range input instead of seconds.
6. Fix the tests that were forced to have wrong expectations because of
the existing bugs.
7. Add missing test coverage for min/max buttons and prop syncing.

# Tests

Unit tests added and updated.
Also tested manually.

# Todos

- [x] Add entry to changelog (if necessary).