Remove flowLessCopy prop from Hash component (#3293)
# Motivation

The `Hash` component can display a copy icon.
This icon is given some extra padding to make sure the touch target is
large enough.
In some contexts this causes more space between lines so we positioned
the icon absolutely.
This required special treatment in those cases and an extra prop to
indicate whether to use the special treatment.

# Changes

1. Change the styling by wrapping the icon in a 0-height inline
component, which works in both cases.
2. Remove the `flowLessCopy` prop which is now no longer necessary.

# Tests

Tried manually on desktop and mobile.
As far as I can tell, the icons that used to have `flowLessCopy` are in
exactly the same place while the other icons moved down 1 pixel. I
believe they are actually more correct now but it's hard to tell because
the icon is 32px while the container is 34.28px.

See: https://fswcu-uqaaa-aaaaa-aadwq-cai.nnsdapp.dfinity.network/

# Todos

- [x] Add entry to changelog (if necessary).