Snsdemo release (#3291)
# Motivation
The snsdemo version is currently specified as a commit in `dfx.json`.
Unfortunately this version is not applied to the snsdemo snapshot.

# Changes
- Use a release tag to specify which version the nns-dapp should use,
instead of a commit.
- Use the release tag to get the sns snapshot.

# Tests
See CI.

# Todos

- [ ] Add entry to changelog (if necessary).
I think this is too minor a tweak to warrant a chnagelog entry. No
frontend or canister changes and only a minor operations tweak.

Co-authored-by: Max Murphy <max@winning.black>