feat: highlight code in new block of Sns aggregator and style (#3300)
# Motivation

Follow-up of #3295.

This PR propose a code highlighter for the "block of code" displayed in
the Sns aggregator index web page and style the new section.

It also adds a sentence to highlight the fact that the aggregator is
meant for developers.

## Library

Prismjs used to highlight the code is fetched from the [Juno's
CDN](https://github.com/buildwithjuno/cdn) on the raw domain of the IC.

# Screenshot

<img width="1536" alt="Capture d’écran 2023-09-12 à 07 06 01"
src="https://github.com/dfinity/nns-dapp/assets/16886711/cc2bb20e-0e71-40a8-9c37-d36088e44d42">