Feat: Add metadata logo path to SNS aggregator data (#3301)
# Motivation

The logo to the SNS Metadata needs to be hardcoded by the frontend
following a specific pattern.

In this PR, I add the relative path of the logo so that the logic of
where the logo lives lies only in the SNS aggregator canister.

Another reason is because the logo is optional. Therefore, this way it's
easier to know whether the logo is present or not. Otherwise, we would
need to try to fetch the asset and check it.

# Changes

* New field `logo` in `SlowMetadata`.
* Change `SlowMetadata::from` for `from_get_metadata_response` because
it uses a second argument for the logo url.

# Tests

Tests are skipped.

# Todos

- [x] Add entry to changelog (if necessary).