Fix: Partial different derived state in snsQueryStore (#3312)
# Motivation

Part of the response of `get_state` is the derived state, which we can
also get from `get_derived_state`. Yet, the endpoint returns different
fields and different precision for one of the fields
(`sns_tokens_per_icp`).

When we were populating the snsQueryStore in the service
`loadSnsProjects`, we were using first the data from the `get_state`,
and right after we were updating it with the data from
`get_derived_state`.

This caused that the sns summary from snsQueryStore and from
snsAggregatorStore didn't match for some time.

# Changes

* Use the data from `get_derived_state` in the first place instead of
updating it afterward.

# Tests

* The tests didn't cover the moment when the stores were different
because it happens in the middle of the function. Yet, when I removed
the call to `updateDerivedState`, the test failed. Therefore, it was
checking that both derived states from snsAggregator and snsQueryStore
where the same. Then, I fixed the test by changing the value in the
field `derived` for the snsQueryStore.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.