Update downgrade-upgrade test (#3315)
# Motivation
The downgrade-upgrade test needs to be able to verify that accounts are
preserved across schema updates.

# Changes
- Make it possible to specify a wasm to downgrade to. This way the test
can be run with a wasm that implements the future schema `S0.early` and
another, `undefined.late` that is not yet in production.
- Get account data; this was already present but commented out as prod
did not have account data in a stable, usable form.

# Tests

<!-- Please provide any information or screenshots about the tests that
have been done -->

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>