GIX-1889: Fix blank modal disburse maturity (#3317)
# Motivation

There was a bug when NeuronSelectPercentage was used for maturity above
1000.

That's because now we started calculating the maturity of the
percentage. Yet, the maturity was passed as a string and then converted
to a number. If the number was larger than 1000, then the string
couldn't be converted successfully to a number.

The solution was to expect a number instead of a string.

# Changes

* Change `formttedMaturity` in NeuronSelectPercentage for
`availableMaturityE8s`. Change accordingly.
* Change where the NeuronSelectPercentage was used to pass the maturity
as e8s instead of as a string.
* Remove unnecessary util `maturityPercentageToE8s`. Now it's a simple
multiplication and division. No need to use an util for that. There is
no logic to reuse.

# Tests

* I added a test that fails in `main` but succeeds in this PR.

# Todos

- [ ] Add entry to changelog (if necessary).