Add feature flag ENABLE_STAKE_NEURON_ICRC1 (#3319)
# Motivation

We want to start using ICRC-1 for the ledger transfers on staking a
neuron.
We want to test it before making it default.

# Changes

Add feature flag ENABLE_STAKE_NEURON_ICRC1. In a later PR this will be
used to enable using ICRC-1 ledger transfers on staking a neuron.

# Tests

no

# Todos

- [ ] Add entry to changelog (if necessary).
I will add an entry when I actually use the flag.