Fix: Docker build (#3333)
# Motivation

docker-build script isn't working.

The problem seems to be when installing ic-wasm with cargo-binstall.

There was a new release of cargo-binstall and this one started breaking
the build script.

In this PR, I install the previous version of cargo-binstall manually
instead of relying in the installer script.

# Changes

* In Dockerfile, install a specific version of cargo-binstall manually.

# Tests

Tested in the gix machine.

# Todos

- [x] Add entry to changelog (if necessary).