Fix: SnsStakeMaturityButton spec file (#3335)
# Motivation

One test for SnsStakeMaturityButton was not correct. The test passed
because the expected text (title of the modal) was the same text as the
button.

In this PR, I added a testId to the modal and checked the presence of
the modal with test id. The test failed.

The test also failed because it required a token in the store to be
present. That is not necessary for the modal, but it was inside an `if`
checking the presence of the token.

Therefore, in this PR, I also refactored the SnsNeuronModals to check
for presence next to the type of the modal. This way it's easier to
understand what is strictly necessary in each modal.

# Changes

* Add testId to StakeMaturityModal.
* Move modals and conditions in SnsNeuronModals.

# Tests

* Use `SnsNeuronContextTest` in `SnsStakeMaturityButton.spec` to check
that modal is opened.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.