Update more aggregator data quickly (#3320)
# Motivation
The swap canister `get_state` API is deprecated and is being removed
from the front end. This is the data that updates rapidly during an SNS
sale. The relevant data will be obtained from `get_derived_state`
instead. Until the front end has finished transitioning, the aggregator
needs to update both datasets frequently during a sale. After the front
end has transitioned, the deprecated `get_state` can be dropped.

# Changes
* Factor out the `get_swap_state` and `get_derived_state` API calls so
that we are guaranteed to get the data in the same way for slow and fast
updates. The API call is not complicated but I want to be 100% sure that
there is no variation.
* Where previously the fast scheduler updated only the swap state, now
update the derived state as well.

# Tests
### Tests in CI
An e2e test is included that creates an SNS and verifies that sample
fast values are updated promptly.

### Tests on Mainnet
You can watch updates to the ICP committed to the SEERS SNS with:
```
while true; do
  date
  printf "%-30s " "Production:"
  curl -s "https://3r4gx-wqaaa-aaaaq-aaaia-cai.raw.icp0.io/v1/sns/root/u67kc-jyaaa-aaaaq-aabpq-cai/slow.json?timestamp=$(date -u +"%Y-%m-%dT%H:%M:%SZ")" | jq '.derived_state.buyer_total_icp_e8s'
  printf "%-30s " "Test deployment on mainnet:"
  curl -s "https://otgyv-wyaaa-aaaak-qcgba-cai.raw.icp0.io/v1/sns/root/u67kc-jyaaa-aaaaq-aabpq-cai/slow.json?timestamp=$(date -u +"%Y-%m-%dT%H:%M:%SZ")" | jq '.derived_state.buyer_total_icp_e8s'
  printf "%-30s " "Swap canister:"
  dfx canister call rmg5p-zaaaa-aaaaq-aabra-cai get_derived_state --network ic --query --candid ./declarations/sns_swap/sns_swap.did '(record {})' 2> /dev/null | idl2json | jq -r .buyer_total_icp_e8s[0]
  sleep 30
done
```
A log of the output is included as a comment below.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>