Fix: Tooltip is hidden if container has overflow hidden (#3339)
# Motivation

Part of the tooltip of the disabled "Disburse" button is hidden behind
the island.

That's because the tooltip uses the window to calculate whether there is
enough space to the left or right. The problem happens when the
container where the tooltip lives has `overflow: hidden`.

The tooltip already received a containerSelector. Therefore, it can use
the container to calculate the space to the left and right.

# Changes

* Use the container element instead of the width to calculate the space
in the left or right of the tooltip.

# Tests

* Purely CSS changes.

# Todos

- [ ] Add entry to changelog (if necessary).