Fix: Better min and max range in disburse maturity (#3340)
# Motivation

The range of the predicted tokens in the disbursing flow is an
approximation of two decimals. By default, the decimals were rounded the
standard way. That meant that the min range could be rounded up and give
a wrong min, or the max range rounded down and give a wrong max.

In this PR, we add the `roundingMode` parameter to `formatToken` and use
it for the range.

# Changes

* Add `roundingMode` param to `formatToken`.
* Use param `roundingMode` for the range in DisburseMaturityModal
confirmation screen.

# Tests

* Add skipped tests because the new parameter is only supported in node
19 and we are in node 18.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary. Covered by an entry.