E2E: Work around flakiness loading proposals (#3342)
# Motivation

We've seen flakiness with `e2e/neurons.spec.ts` for a while. I think I
finally figured out what happens.

The `NnsProposals` component loads neurons and proposals at the same
time.
But once neurons are loaded, it loads proposals again. So it's possible
that the component goes back into loading state immediately after
proposals are loaded.

I was able to reproduce this by adding an artificial delay in
`listNeurons`. Then you consistently see the skeletons reappear after
the disappear.

Unfortunately the component is very complicated so it's not easy to fix
without a major cleanup.
So until we have time to clean this up, we should work around the issue
to avoid flakiness with the test.

# Changes

Once proposals are loaded, wait 1 second and wait again for proposals to
be loaded if necessary.

# Tests

Still pass. Hopefully less flaky in the future.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary