Add sendIcpIcrc1 to ICP ledger API (#3349)
# Motivation

We want to use the ICRC-1 interface of the ICP ledger canister when
participating in a swap or topping up a canister etc.

# Changes

Add `sendIcpIcrc1` function to `icp-ledger.api` similar to the existing
`sendICP` function on the same API.

# Tests

Duplicated the existing tests for `sendICP` and added a test for the fee
that can be passed in.

# Todos

- [ ] Add entry to changelog (if necessary).
Will add when it's used.