Include the schema in the stats (#3323)
# Motivation
When changing the schema, a migration completes at some time after the
deployment of a new canister. It is necessary to have visibility of when
that migration completes.

The most important information is what the canonical storage schema is.
Some sort of progress-meter would be nice but not essential and depends
rather on the details of a specific migration. A countdown to zero is
reasonably general and the meaning of the count can be specified in a
different way for every migration.

# Changes
* Add the current schema to the metrics.

# Tests
The metric only becomes interesting when we have migrations. We can
however verify that artificial changes in the schema are passed through
to the metrics and API:

Artificial changes:
```
 pub enum SchemaLabel {
-    Map = 0,
+    Map = 1,
 }
 
     pub fn migration_countdown(&self) -> u32 {
-        0
+        1234
     }
```
This yields this metrics:
```
# HELP nns_dapp_schema The nns-dapp schema version
# TYPE nns_dapp_schema gauge
nns_dapp_schema 1 1694800609059
# HELP nns_dapp_migration_countdown When non-zero, a migration is in progress.
# TYPE nns_dapp_migration_countdown gauge
nns_dapp_migration_countdown 1234 1694800609059
```
and stats:
```
max@sinkpad:~/dfn/nns-dapp/branches/migration-status (1:46:10)$ dfx canister call nns-dapp get_stats
(
  record {
    migration_countdown = opt (1_234 : nat32);
   [SNIP]
    schema = opt (1 : nat32);
```

# Todos

- [ ] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>