Remove extra space from AmountDisplay (#3391)
# Motivation

https://github.com/dfinity/nns-dapp/pull/3379 had to add a `trim()`
which shouldn't be necessary.
It was necessary because `AmountDisplay` renders an unnecessary space.
`AmountDisplay` has an optional copy icon. If the copy icon is rendered,
we want there to be a space between the amount and the copy icon. But if
there is no copy icon, there also shouldn't be a space.
I think this might be a bug in Svelte template parsing but we can work
around it.

# Changes

1. Render the space separating the amount and the copy icon inside the
`{#if copy}` block.
2. Remove the `trim()` from
`frontend/src/tests/page-objects/KeyValuePair.page-object.ts`

# Tests

Pass
Also manually checked that it renders correctly with copy icon (in the
wallet summary) and without (in the transaction history).
Try it here:
https://fswcu-uqaaa-aaaaa-aadwq-cai.nnsdapp.dfinity.network/

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary