Chore: Update next dependencies (#3404)
# Motivation

We want to use the new fields related to the NF enhancements for swaps.

In this PR, I updated the next dependencies and fixed some
incompatibilities.

# Changes

## Automatic changes

Running `npm run update:next`.
* Changes in package-lock.json.

## Manual

There were broken changes because the aggregator utils needed to return
the new fields also. An empty array is not an optional field in JS...
* Changes in
`frontend/src/lib/utils/sns-aggregator-converters.utils.ts`.
The support for this fields in the SNS aggretor is in `main`. But I will
add the integration with those fields in another PR.
* All changes in tests.

# Tests

* Add new fields to all mocks and expects.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet necessary.