GIX-1930: Use new fields from sns aggregator (#3411)
# Motivation

We want to expose the new fields related to NF and already in the SNS
Aggregator response to the nns-dapp components.

In this PR, the converters of the SNS aggregator data use the fields in
the response to populate the converted data, instead of returning empty
arrays.

# Changes

* Add the new fields to the "CachedXxx" types.
* Create new CachedNeuronsFundParticipationConstraints and
CachedLinearScalingCoefficient types.
* New converted `convertNeuronsFundParticipationConstraints`.
* Use the new fields in the relevant converters.

# Tests

* Test that an aggregator response with the new fields, includes them in
the SnsSummary.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary yet.