Feat: Change canister detail page title (#3424)
# Motivation

The title "Balance unavailable" was not very user friendly when the user
was not the controller.

Instead, the title will be now the canister name or the id when the user
is not the controller. In that case, there is no need for the subtitle.

# Changes

* Change the title in CanisterHeadingTitle when the user is not the
controller. To access the name, I added the prop `canister`.
* Pass the `canister` prop to CanisterHeadingTitle in
CanisterPageHeading.
* Add the logic of not the controller to render the PageHeading without
title in CanisterPageHeading.
* Remove the `balance_unavailable` i18n key.

# Tests

* Change and add test cases for new functionality in
CanisterHeadingTItle.spec
* Change and add test cases for new functionality in
CanisterPageHeading.spec
* Change and add test cases for new functionality in CanisterDetail.spec

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary. Covered by the entry that there is a new canister header.