Extend release-sop script (#3418)
# Motivation

Currently, the `release-sop` script only enacts the release checklist
until the point where the RC is deployed to staging.
We want the `release-sop` script to guide the entire release process.
This PR takes the script up to submitting the proposal.
Continuing the successful release is for a future PR.

# Changes

1. Extend the `release-sop` script up to the point where the proposal is
submitted.
2. Update the `release` script with a flag to store the proposal ID of
the created proposal.
3. Update tests of `release-sop`.

# Tests

I did this week's release with the new script.

# Todos

- [x] Add entry to changelog (if necessary).